+++ universal-login-wallet/src/ui/App.tsx
@@ -9,7 +9,7 @@ import NotificationsScreen from './Notifications/NotificationsScreen';
 import ApproveScreen from './Login/ApproveScreen';
 import RecoveryScreen from './Login/RecoveryScreen';
 import SettingsScreen from './Settings/SettingsScreen';
-import {useServices} from '../hooks';
+import {useServices} from './hooks';
 
 const App = () => {
   const {walletService} = useServices();
