+++ universal-login-wallet/src/ui/Home/Balance.tsx
@@ -1,5 +1,5 @@
 import React, {useState, useEffect} from 'react';
-import {useServices} from '../../hooks';
+import {useServices} from '../hooks';
 import {utils} from 'ethers';
 import EthereumIcon from './../assets/icons/ethereum.svg';
