+++ universal-login-wallet/src/ui/Modals/ModalPersonalInfo.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useServices } from './../../hooks/useServices';
+import { useServices } from '../hooks/useServices';
 import Input from '../common/Input';
 import InputLabel from '../common/InputLabel';
 import ButtonFullwidth from '../common/ButtonFullwidth';
