+++ universal-login-wallet/src/ui/Modals/ModalSafello.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {getSafelloUrl} from '../utils';
-import {useServices} from '../../hooks';
+import {useServices} from '../hooks';
 
 
 const ModalSafello = () => {
