+++ universal-login-wallet/src/ui/Modals/ModalTopUp.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useServices } from './../../hooks/useServices';
+import { useServices } from '../hooks/useServices';
 import usdIcon from '../assets/icons/usd.png';
 import etherIcon from '../assets/icons/ether.svg';
 import bitmapIcon from '../assets/icons/bitmap.svg';
