+++ universal-login-wallet/src/ui/Modals/ModalTransfer.tsx
@@ -3,7 +3,7 @@ import Input from '../common/Input';
 import InputLabel from '../common/InputLabel';
 import InputWithDropdown from '../common/InputWithDropdown';
 import ButtonFullwidth from '../common/ButtonFullwidth';
-import {useServices} from '../../hooks';
+import {useServices} from '../hooks';
 import {TransferDetails} from '../../services/TransferService';
