+++ universal-login-wallet/src/ui/common/Sidebar.tsx
@@ -2,7 +2,7 @@ import React, {useEffect, useState} from 'react';
 import logo from '../assets/logo.svg';
 import { NavLink } from 'react-router-dom';
 import avatar from '../assets/avatar.svg';
-import { useServices } from '../../hooks';
+import { useServices } from '../hooks';
 
 function Sidebar() {
   const {userDropdownService, notificationService} = useServices();
