+++ universal-login-wallet/src/ui/common/UserDropdown.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import avatar from './../assets/avatar.svg';
-import {useRouter, useServices, useSubscription} from '../../hooks';
+import {useRouter, useServices, useSubscription} from '../hooks';
 
 
 const UserDropdown = () => {
