+++ universal-login-wallet/src/ui/hooks/subscription.ts
@@ -1,5 +1,5 @@
 import {useState, useEffect} from 'react';
-import {Subject} from '../services/Subject';
+import {Subject} from '../../services/Subject';
 
 export function useSubscription<T>(subject: Subject<T>) {
   const [event, setEvent] = useState<T | undefined>(undefined);
