+++ universal-login-wallet/src/ui/hooks/useRouter.ts
@@ -1,5 +1,5 @@
 import {useContext} from 'react';
-import {RouterContext} from '../ui/CustomBrowserRouter';
+import {RouterContext} from '../CustomBrowserRouter';
 
 export function useRouter() {
   return useContext(RouterContext);
