+++ universal-login-wallet/src/ui/hooks/useServices.ts
@@ -1,5 +1,5 @@
 import {useContext} from 'react';
-import {ServiceContext} from '../services/Services';
+import {ServiceContext} from '../../services/Services';
 
 export function useServices() {
   return useContext(ServiceContext);
